+++ retools/tests/test_util.py
@@ -20,7 +20,7 @@ class TestNamespaceFunc(unittest.TestCase):
 class TestContextManager(unittest.TestCase):
     def _call_with_contexts(self, ctx_managers, func, kwargs):
         from retools.util import with_nested_contexts
-        return with_nested_contexts(ctx_managers, func, kwargs)
+        return with_nested_contexts(ctx_managers, func, [], kwargs)
     
     def test_nest_call(self):
         def a_func(**kwargs):
@@ -28,14 +28,14 @@ class TestContextManager(unittest.TestCase):
             return kwargs['list']
         
         @contextmanager
-        def ctx_a(func, **kwargs):
+        def ctx_a(func, *args, **kwargs):
             eq_(func, a_func)
             kwargs['list'].append(0)
             yield
             kwargs['list'].append(1)
         
         @contextmanager
-        def ctx_b(func, **kwargs):
+        def ctx_b(func, *args, **kwargs):
             eq_(func, a_func)
             kwargs['list'].append(2)
             yield